Factor out dissimilarity heatmap config into shared file in anticipation of plot_canon_heatmap
