+++ setup.cfg
@@ -22,6 +22,8 @@ ignore = W503,E203
 known_first_party=evaluating_rewards,tests
 default_section=THIRDPARTY
 multi_line_output=3
+include_trailing_comma=True
+use_parentheses=True
 force_sort_within_sections=True
 line_length=100
